requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=h27087fc_1
aiohttp=3.8.1=py39hb9d737c_1
aiosignal=1.3.1=pyhd8ed1ab_0
appdirs=1.4.4=pyh9f0ad1d_0
arrow-cpp=8.0.0=py39h60b952e_1
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.2.0=pyh71513ae_0
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
blas=1.0=mkl
boost-cpp=1.78.0=he72f1d9_0
bottleneck=1.3.5=py39h7deecbd_0
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=unix_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
cryptography=39.0.1=py39h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
dataclasses=0.8=pyhc8e2a94_3
datasets=2.11.0=py_0
dill=0.3.6=pyhd8ed1ab_1
docker-pycreds=0.4.0=py_0
docopt=0.6.2=py39h06a4308_1
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
filelock=3.10.7=pyhd8ed1ab_0
flit-core=3.8.0=py39h06a4308_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py39h5eee18b_0
fsspec=2023.3.0=pyhd8ed1ab_1
gflags=2.2.2=he1b5a44_1004
giflib=5.2.1=h5eee18b_3
gitdb=4.0.10=pyhd8ed1ab_0
gitpython=3.1.31=pyhd8ed1ab_0
glog=0.6.0=h6f12383_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.46.1=h33aed49_1
huggingface_hub=0.13.3=pyhd8ed1ab_0
icu=70.1=h27087fc_0
idna=3.4=py39h06a4308_0
importlib-metadata=6.1.0=pyha770c72_0
importlib_metadata=6.1.0=hd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h5eee18b_1
keyutils=1.6.1=h166bdaf_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcurl=7.87.0=h91b91d3_0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libthrift=0.15.0=he6d91bd_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuv=1.44.2=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=6.0.2=py39h5eee18b_0
multiprocess=0.70.14=py39h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
numexpr=2.8.4=py39he184ba9_0
numpy=1.23.5=py39h14f4228_0
numpy-base=1.23.5=py39h31eccc5_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
orc=1.7.4=hb3bc3d3_1
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.2=py39h417a72b_0
pathtools=0.1.2=py_1
pillow=9.4.0=py39h6a678d5_0
pip=23.0.1=py39h06a4308_0
protobuf=3.20.3=py39h6a678d5_0
psutil=5.9.0=py39h5eee18b_0
pyarrow=8.0.0=py39h992f0b0_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.0.0=py39h06a4308_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.15=h7a1cb2a_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python-xxhash=3.0.0=py39hb9d737c_1
python_abi=3.9=2_cp39
pytorch=1.11.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2023.3=pyhd8ed1ab_0
pyyaml=6.0=py39hb9d737c_4
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
regex=2022.7.9=py39h5eee18b_0
requests=2.28.1=py39h06a4308_1
responses=0.18.0=pyhd8ed1ab_0
sacremoses=0.0.53=pyhd8ed1ab_0
scikit-learn=1.2.0=py39h6a678d5_1
scipy=1.9.3=py39h14f4228_0
sentencepiece=0.1.95=py39hd09550d_0
sentry-sdk=1.18.0=pyhd8ed1ab_0
setproctitle=1.2.2=py39hb9d737c_2
setuptools=65.6.3=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=3.0.5=pyh44b312d_0
snappy=1.1.9=h295c915_0
sqlite=3.41.1=h5eee18b_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.11.4=py39h3dcd8bd_1
torchaudio=0.11.0=py39_cu113
torchvision=0.12.0=py39_cu113
tqdm=4.65.0=pyhd8ed1ab_1
transformers=4.27.4=pyhd8ed1ab_0
typing-extensions=4.4.0=py39h06a4308_0
typing_extensions=4.4.0=py39h06a4308_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.14=py39h06a4308_0
utf8proc=2.6.1=h27cfd23_0
wandb=0.14.0=pyhd8ed1ab_0
wheel=0.38.4=py39h06a4308_0
xxhash=0.8.0=h7f98852_3
xz=5.2.10=h5eee18b_1
yaml=0.2.5=h7f98852_2
yarl=1.7.2=py39hb9d737c_2
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.4=hc292b87_0
